Modified spotter linux makefile to use cleaner fs
Updated SimplePhraseSpotter to use a cleaner directory expectation
for the files it borrows from HoundifyExplorer. I wanted to get
this handled now so that when I write the documentation it
reflects a cleaner working space.

Previously this make file required that the files borrowed from
HoundifyExplorer be just dumped in the root of a local copy of
this repository' now those files must be in a directory named
HoundifyExplorerUtils in the root of a local copy of this repo.