Renamed directories
I've renamed the SimpleAudioQueriesOnlyWrittenResponseAndNoHistory
directory to be SimpleAudioQueriesOnlyWrittenResponse.

I've made this change because it turns out that the HoundConverser
will manage recording the conversation state and passing it
between queries for us so their's no need for us to do it
ourselves. And it saves me from needing to write another simple
driver to illustrate that.