Added a recommended hardware section to README.md
Added a recommended hardware section to the README.md.  This will
be hardware that will recieve first class support by this repository.

Note, the Simple Drivers will try to continue to work with the Linux
and the Pi 3 B+ with any ALSA compatible microphone device and
ALSA compatible speaker device; however the files in the main directory
will be more focused on the recommended hardware.