Added new folder structure for SmartDevices
I added a new folder structure that will contain code that can be
used with different smart devices. SmartDevices/, along with
SmartDevices/lib/ and all its subdirectories, will have a
.gitignore file that ignores all .o files. While,
SmartDevices/include and all its subdirectories will include a
.gitignore file that ignores all non .h files.

I also added a SmartPlugs directory to both lib and include as
the first SmartDevice code I'll be committing will be for a
SmartPlug!