Added Respeaker LED support to SimpleAudioQueriesOnlyWrittenResponse
I added support for the ReSpeaker 4 Mic Array's LED ring to the
SimpleAudioQueriesOnlyWrittenResponse driver.  This support is guarded
by RESPEAKERLEDRING macros. This is to allow this simple driver to
continue to work even if the user doesn't want to deal with the extra
dependencies introduced by this feature.  The two major dependancy
additions here are: the use of the WiringPi library and OpenMP.

WiringPi is used to communicate with the GPIO port on the Pi.

OpenMP is used to create two sperate threads after the microphone
hears "Ok Hound". The first thread is responsible for spining
the LEDs while the second thread is responsible for sending the
audio data to the server and telling the first thred when to stop
spinning the LED ring. (Which it does through an object of the
SimpleSinkTurnOffLight class)