Fixed DEFNINE spelling mistake and add DEFNINE_DEBUG_FLAGS
I fixed a spelling mistake issue that revolved around the use of DEFNINE
instead of DEFINE.

I also added support for DEFINE_FEATURE_FLAGS to allow users to overide
this variable with flags that hey would like to pass to enable different
features.

Currently the only expected flag is -DRESPEAKERLEDRING and in addition
to passing this flag onward if -DRESPEAKERLEDRING is detected then
additional options are passed at different stages. Since -DRESPEAKERLEDRING
depends on OpenMP and WiringPi the flags for them are added to the
linking step if -DRESPEAKERLEDRING is detected in DEFINE_FEATURE_FLAGS
the flags added at linking are:

-lwiringPi -lwiringPiDev -lm -lcrypt -lrt -lgomp -fopenmp

Additionally the -fopenmp is also included during compilation of
SimpleAudioQueriesOnlyWrittenResponse.o and LocalConfig.o.

An example call of make that uses this feature would be:

    make -f Makefile.Rpi DEFINE_FEATURE_FLAGS='-DRESPEAKERLEDRING'