Added --showstartup to SimpleAudioQueries
I added a --showstartup flag to the SimpleAudioQueries driver
when this driver is compiled with respeaker led features.

When this flag is used the respeaker leds will light up green
for 5 seconds once the device has started listening for the
wakephrase.  Users can interrupt this by saying the wake
phrase and they do not need to wait for the LEDs to turn off.

I accomplished this by using an std::thread with an
std::atomic<bool> that detatchs and then after 5 seconds
checks if it should still turn off the LEDs.  If so it does.