Switched to using lambda instead of static function
I updated the simple audio queries driver to use a lambda instead
of a static function for the time out thread.

I did this to simplify the code since this function is only
intended to be used in a single place it makes more sense to in
line it using a lambda.