Corrected some unintended modifications of mkfile
I corrected some unintended modifcations to the SimpleAudioQueries
RPI makefile. A lot of changes have been happening to this file
in the master branch and when syncing with master this branch
got some unintended modifications.  The modifications that occured
were losing the include and linking steps for the SmartDevices
files.