Add --num-of-mic-channels to simpleaudioqueries
I added --num-of-mic-channels flag to the SimpleAudioQueries
driver.  This flag passes the number of channels to use when
transmitting data to the houndify cloud. Note the OkHound
phrase spotter only supports single channel so the value of this
flag is ignored for the wake phrase.  If --num-of-mic-channels
is not passed then 1 is assumed.