Added some simple headers
These simple headers are SimplePartialHandler.h and
SimpleRequestInfoPreparer.h.

SimplePartialHandler.h contains a class named SimplePartialHandler
that can be used as a partial handler in a voice request to the
Houndify platform.  It allows users to see the partial transcript
updated as the partial responses return from the server.  If
HOUNDHOUSEDEBUG is defined then every partial transcript is
printed on its own line.

SimpleRequestInfoPreparer.h contains a class named
SimpleRequestInfoPreparer that can be used with different classes
from the Houndify SDK to prepare a RequestInfo. If HOUNDHOUSEDEBUG
is defined then a debug message is printed when preparations
begin and when they end.

These simple headers will be used in a new simple driver that I
will be adding later.