Added SimpleSinks.h
I've added a new simple header file called SimpleSinks.h this file
will be use for the declaration of simple sinks that are useful
to use within this repository and elsewhere.

All classes in this file should be subclasses of
ClientCapabilityRegistry::DataSink contained in
ClientCapabilityRegistry.h within the HoundCpp SDK.

At the time of adding this file there's only one sink within it.
That sink is SimpleSinkWithoutClosingSound and has two features
worth highlighting. First, it will continue passing audio to
the voice_request as long as more_ok returns true and the server
hasn't told the partial handler that it thinks it should stop
sending data. Second, it would normally be appropriate to play
the closing sound bite when write_data returns 0. However, this
class is deliberately meant for situations that don't use audio
output, so it does not do so.