Added files to make a simple phrase spotter
Commit includes a cpp file that compiles to a simple driver
program that serves as an example phrase spotter.

Additionally I'm checking in a Makefile that can be used to
compile this driver on a Raspberry Pi 3 b+.  Note, this Makefile
uses files from HoundCpp and files from HoundifyExplorer.

README.md already explains how to prep HoundCpp for use by this
Makefile and in a different commit I'll update it to include
instructions on how to extract the needed files for this Makefile.