Updated README.md to include HoundifyExplorer
Added a step to the getting started secction that explains how to
extract some useful utilities from HoundifyExplorer. For
now only files that are useful to the SimplePhraseSpotter are
extracted; however there are many more useful utlities in
HoundifyExplorer so as this client's use of those utilities grows
so will this section.