Adding a Makefile.Linux for new simple driver
Adding a make file for a new simple driver that will be named
SimpleAudioQueriesOnlyWrittenResponseAndNoHistory, this simple
driver will support the user saying "Okay Hound" followed by
a query. Additionally the user will see their transcript appear on
the screen as the server tries to figure out what they said. Then
finally the user will see the response that was generated by the
server in its written form only.

One other thing to note about this driver is that it will not
support passing the conversations state from one query to the
next. So it will be as if the conversation starts fresh with every
query.