Added a basic config mechanism
I've added a new simple config mechanism that will allow users
to specify thier own client id, client key, and user id and allows
developers to work on this repository without revealing this
private information.

To use this a user must include a definition that matches this
declaration: const SimpleHoundCloudRequesterConfig simple_config

I've included a LocalConfig.cpp.example that users can cp to
LocalConfig.cpp and use after changing the values where appropriate.

Note, this mechanism will probably change to something like a
config file or command line params in the near future because it
takes a long time to compile on the Pi 3 B+ and it's not necessary
to hard code these values and force a recompile every time they
change.

I've also updated the linux makefile for the new simple driver
to use this mechanism.