Added a new simple audio verifier
I've added a new simple audio verifier, this verifier is named
SimpleAudioVerifier and is as subclass of HoundAudioVerifier.
HoundAudioVerifier is a class that comes from HoundifyExplorer
so I've also added a new command to the README.md that shows how
to bring that file over for use by this repo.