Added a new header for dynamic resp handlers
Added a new simple header file named
SimpleDynamicResponseHandlers.h and is meant to contain the
declarations of simple subclasses of the DynamicResponseHandler
class within the HoundCpp SDK.

At time off adding SimpleDynamicResponseHandlerWithNoAudioOutput
is the only subclass currently in this file. More will be added
later but this basic one will be used in the new driver. It
doesn't do anything special just prints the written response
to std::cout.